+++ material/src/material.ts
@@ -1,4 +1,5 @@
 export * from './modules/dialogPopup.module';
+export * from './modules/virtualEditPopup.module';
 export * from './components/basicDialogPopup/types';
 export * from './plugins/popup/components';
 export * from './plugins/popup';