+++ material/src/plugins/popup/components.ts
@@ -1,1 +1,2 @@
 export * from './dialog/dialogPopup.component';
+export * from './virtualEdit/virtualEditPopup.component';