+++ material/src/plugins/popup/index.ts
@@ -1,1 +1,2 @@
 export * from './dialog/dialogPopup.interface';
+export * from './virtualEdit/virtualEditPopup.interface';