+++ material/src/plugins/popup/virtualEdit/virtualEditPopup.interface.ts
@@ -1,0 +1,22 @@
+import {Popup, PopupOptions} from "@anglr/select";
+
+/**
+ * Css classes for virtual edit popup
+ */
+export interface CssClassesVirtualEditPopup
+{
+}
+
+/**
+ * Virtual edit popup options
+ */
+export interface VirtualEditPopupOptions extends PopupOptions<CssClassesVirtualEditPopup>
+{
+}
+
+/**
+ * Public API for 'VirtualEditPopupComponent'
+ */
+export interface VirtualEditPopup extends Popup
+{
+}